+++ src/ts/Views/View.ts
@@ -82,5 +82,4 @@ export abstract class View<T extends object> {
     public hide(): void {
         this._container.style.visibility = 'hidden';
     }
-
 }